client: Merge iota_api into client as extended API
