+++ iota-client/extended/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod get_inputs;
+pub(crate) mod get_new_address;
+pub(crate) mod prepare_transfers;
+pub(crate) mod send_transfers;
+pub(crate) mod send_trytes;
