+++ src/prelude.rs
@@ -1,5 +1,5 @@
 pub use crate::{
-    iota_api, iota_client, iota_constants, iota_conversion, iota_crypto, iota_model, iota_pow,
+    iota_client, iota_constants, iota_conversion, iota_crypto, iota_model, iota_pow,
     iota_signing, iota_utils, iota_validation,
 };
 /// This trait provides conversion methods for various types into trinary
